+++ src/solver/task/tracing/tracer.rs
@@ -30,11 +30,6 @@ pub trait Tracer: Send + Sync {
 
     // Called on hypothesis processing finished
     fn on_hypothesis_finish(&mut self, _inference: &TermInference, _cycle: usize) {}
-
-    // Cancel task form outside
-    fn is_cancelled(&self) -> bool {
-        false
-    }
 }
 
 #[derive(Default)]
@@ -101,9 +96,5 @@ impl Tracer for TracerHub {
         for i in self.tracers.iter_mut() {
             i.on_hypothesis_finish(inference, cycle);
         }
-    }
-
-    fn is_cancelled(&self) -> bool {
-        self.tracers.iter().any(|x| x.is_cancelled())
     }
 }